+++ packages/implementations/react/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 `@rarui-react/components` components is a component library built with [React](https://reactjs.org).
 
+## 2024-05-15 `1.19.0`
+
+#### 🎉 New features
+
+- Added new `backgroundPositionX` and `backgroundPositionY` properties to the component `Box`. ([#38](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/38) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new atomic `Sidebar` component. ([#38](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/38) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new atomic `SideNavigation` component. ([#38](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/38) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-05-07 `1.18.0`
 
 #### 🎉 New features